why did she tell us to change our table names to uppercase this has broken everything

+++ src/main/java/ca/kpu/info2413/library/backend/model/AccountPublication.java
@@ -27,5 +27,5 @@ public class AccountPublication
     private Publication publication;
 
     @Column(name = "waitlist_position")
-    private Integer waitlistPosition;
+    private Short waitlistPosition;
 }
